+++ cmd/podman/main_remote_windows.go
@@ -1,0 +1,7 @@
+// +build remoteclient,windows
+
+package main
+
+func setSyslog() error {
+	return nil
+}
